Simplify enums, use float over decimal, remove unneeded setter

+++ src/Ryujinx.Ava/UI/ViewModels/MainWindowViewModel.cs
@@ -443,7 +443,7 @@ namespace Ryujinx.Ava.UI.ViewModels
             get => _customVSyncIntervalPercentageProxy;
             set
             {
-                int newInterval = (int)(((decimal)value / 100) * 60);
+                int newInterval = (int)((value / 100f) * 60);
                 _customVSyncInterval = newInterval;
                 _customVSyncIntervalPercentageProxy = value;
                 ConfigurationState.Instance.Graphics.CustomVSyncInterval.Value = newInterval;
@@ -476,7 +476,7 @@ namespace Ryujinx.Ava.UI.ViewModels
             set
             {
                 _customVSyncInterval = value;
-                int newPercent = (int)(((decimal)value / 60) * 100);
+                int newPercent = (int)((value / 60f) * 100);
                 _customVSyncIntervalPercentageProxy = newPercent;
                 ConfigurationState.Instance.Graphics.CustomVSyncInterval.Value = value;
                 if (_isGameRunning)
